Conflicts:
	src/main/java/com/Airline/dao/entity/BookingDetails.java
	src/main/java/com/Airline/dao/entity/Inventory.java

+++ src/main/java/com/Airline/dao/entity/AirlineInfo.java
@@ -18,7 +18,7 @@ public class AirlineInfo implements Serializable {
 	@Id
 	@GeneratedValue(strategy=GenerationType.IDENTITY)
 	@Column(name="airline_id")
-	private Integer airlineId;
+	private Long airlineId;
 	
 	@Column(name="airline_logo")
 	private String airlineLogo;
@@ -30,11 +30,11 @@ public class AirlineInfo implements Serializable {
 		
 	}
 
-	public Integer getAirlineId() {
+	public Long getAirlineId() {
 		return airlineId;
 	}
 
-	public void setAirlineId(Integer airlineId) {
+	public void setAirlineId(Long airlineId) {
 		this.airlineId = airlineId;
 	}
