+++ src/main/java/com/Airline/dao/entity/Fare.java
@@ -14,7 +14,7 @@ public class Fare {
 	@Id
 	@GeneratedValue(strategy=GenerationType.IDENTITY)
 	@Column(name="fare_id")
-	private Integer fareId;
+	private Long fareId;
 	
 	@Column(name="currency")
 	private String currency;
@@ -26,11 +26,11 @@ public class Fare {
 		
 	}
 
-	public Integer getFareId() {
+	public Long getFareId() {
 		return fareId;
 	}
 
-	public void setFareId(Integer fareId) {
+	public void setFareId(Long fareId) {
 		this.fareId = fareId;
 	}
