+++ src/main/java/com/Airline/dao/entity/FlightInfo.java
@@ -18,7 +18,7 @@ public class FlightInfo implements Serializable{
 	@Id
 	@GeneratedValue(strategy=GenerationType.IDENTITY)
 	@Column(name="flight_infoid")
-	private Integer flightInfoId;
+	private Long flightInfoId;
 	
 	@Column(name="flight_number")
 	private String flightNumber;
@@ -33,10 +33,10 @@ public class FlightInfo implements Serializable{
 		
 	}
 	
-	public Integer getFlightInfoId() {
+	public Long getFlightInfoId() {
 		return flightInfoId;
 	}
-	public void setFlightInfoId(Integer flightInfoId) {
+	public void setFlightInfoId(Long flightInfoId) {
 		this.flightInfoId = flightInfoId;
 	}
 	public String getFlightNumber() {
